recipe_debug_3

+++ atLaserSoldering/Library/IOLibrary/Class/SolderingPosition.cs
@@ -16,8 +16,8 @@ namespace RecipeManager
         INSPECTION_POSITION_MODE _positiontype = INSPECTION_POSITION_MODE.POSITION_INSPECTION_MODE;
         //INSPECTION_POSITION_MODE _eInspectionPositionMode = INSPECTION_POSITION_MODE.INSPECTION_POSITION_BASE_MODE;
 
-        private int _PreheatPowerRatio = 0;
-        private int _HeatPowerRatio = 0;
+        private double _PreheatPowerRatio = 0;
+        private double _HeatPowerRatio = 0;
         private int _ReadyTime = 0;
         private int _PreHeatTime = 0;
         private int _HeatTime = 0;
@@ -43,8 +43,8 @@ namespace RecipeManager
         public double PositionRx { get { return _positionRx; } set { _positionRx = value; } }
         public double PositionRy { get { return _positionRy; } set { _positionRy = value; } }
         public double PositionRz { get { return _positionRz; } set { _positionRz = value; } }
-        public int PreHeatPowerRatio { get { return _PreheatPowerRatio; } set { _PreheatPowerRatio = value; } }
-        public int HeatPowerRatio { get { return _HeatPowerRatio; } set { _HeatPowerRatio = value; } }
+        public double PreHeatPowerRatio { get { return _PreheatPowerRatio; } set { _PreheatPowerRatio = value; } }
+        public double HeatPowerRatio { get { return _HeatPowerRatio; } set { _HeatPowerRatio = value; } }
         public int ReadyTime { get { return _ReadyTime; } set { _ReadyTime = value; } }
         public int PreHeatTime { get { return _PreHeatTime; } set { _PreHeatTime = value; } }
         public int HeatTime { get { return _HeatTime; } set { _HeatTime = value; } }
